# src/Storybook.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Storybook",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Donation/{donationId}": {
      "get": {
        "tags": [
          "Donation"
        ],
        "summary": "Get Donation by id.",
        "description": "Get Donation by id.",
        "operationId": "getDonationById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "donationId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDonationByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Donation"
        ],
        "summary": "Delete Donation.",
        "description": "Delete Donation.",
        "operationId": "removeDonation",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "donationId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveDonationResponse"
            }
          }
        }
      }
    },
    "/api/Donation": {
      "get": {
        "tags": [
          "Donation"
        ],
        "summary": "Get Donations.",
        "description": "Get Donations.",
        "operationId": "getDonations",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDonationsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Donation"
        ],
        "summary": "Create Donation.",
        "description": "Create Donation.",
        "operationId": "createDonation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateDonationRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateDonationResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Donation"
        ],
        "summary": "Update Donation.",
        "description": "Update Donation.",
        "operationId": "updateDonation",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateDonationRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateDonationResponse"
            }
          }
        }
      }
    },
    "/api/Donation/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "Donation"
        ],
        "summary": "Get Donation Page.",
        "description": "Get Donation Page.",
        "operationId": "getDonationsPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetDonationsPageResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateDonationRequest": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        }
      }
    },
    "CreateDonationResponse": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "DonationDto": {
      "type": "object",
      "properties": {
        "donationId": {
          "format": "int32",
          "type": "integer"
        },
        "amount": {
          "format": "double",
          "type": "number"
        },
        "paymentMethodName": {
          "type": "string"
        },
        "isPerpetual": {
          "type": "boolean"
        }
      }
    },
...
    "GetDonationByIdResponse": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetDonationsPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DonationDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetDonationsResponse": {
      "type": "object",
      "properties": {
        "donations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DonationDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
...
    "RemoveDonationResponse": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "UpdateDonationRequest": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        }
      }
    },
    "UpdateDonationResponse": {
      "type": "object",
      "properties": {
        "donation": {
          "$ref": "#/definitions/DonationDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
  }
}